xdp: widen the two CAS targets to 64-bit to lower the build-clang floor
session_state.alive and sweep.inited are the targets of
__sync_val_compare_and_swap, and the BPF backend below clang 20 has no
32-bit atomic compare-and-swap: it refuses the object at compile with
Unsupported atomic operations, please use 64 bit version, which is the
compiler naming this fix. Both are now __u64. alive is a 0/1 flag and
inited a once-latch, so the width is all that changes; the cost is 4 bytes
per session in the map value, not on the stack.

session_state grows 120 to 128 and the offsets from alive onward shift,
not uniformly: echo_rx_pkts moves +8 not +4 because the u64 realigns. The
abi_check pins were updated from measured offsetof output rather than by
hand, and the two loaders that printed alive with %u were cast.

What this buys, measured, and what it does not: with the widening a stock
Debian 12 clang 14 now COMPILES the object, where before it errored. It
does not yet LOAD what clang 14 produces: on 6.12 that object is refused
with combined stack size of 3 calls is 528. Too large, because clang
14's codegen spills more than clang 21's, which produces 480 and loads.
So the build floor is not clang 14; it is whichever version's codegen
first fits the 512-byte budget, and finding it needs a bisect across
15..20 (Ubuntu 24.04 is 18, Debian 13 is 19). The widening is necessary
for the floor to move at all; it is not sufficient on its own.

Verified on clang 21: make check, and the object loads on the DUT. The
6.1 retest is pending a fresh 6.1 VM, but the widening touches neither the
stack nor the bound pins that 6.1 cares about, so it should be unaffected.

+++ include/bfd_shared.h
@@ -261,9 +261,15 @@ struct session_state {
 	__u8  detect_mult;
 	__u8  remote_flags;  /* peer's last control packet flags,
 	                      * masked to the six non-state bits */
-	__u32 alive;          /* our sweep's verdict: 1 = hearing peer.
-	                       * 32-bit: BPF atomics need 32/64-bit; RX
-	                       * set and sweep clear race across CPUs. */
+	__u64 alive;          /* our sweep's verdict: 1 = hearing peer.
+	                       * 64-bit, not 32: the BPF backend below clang
+	                       * 20 has no 32-bit atomic compare-and-swap and
+	                       * refuses the object with "Unsupported atomic
+	                       * operations, please use 64 bit version". A
+	                       * 64-bit flag lowers the build-clang floor to
+	                       * where distro toolchains sit, at 4 bytes per
+	                       * session. RX set and sweep clear race across
+	                       * CPUs, which is why it is atomic at all. */
 	__u32 final_seq;      /* kernel ack of a Poll sequence: set to
 	                       * cfg->poll_seq on the peer's F */
 	__u8  peer_mac[6];    /* neighbour's source MAC, learned on every RX.
